refactor: move organization class to lib for DRY, refactor climate/index to .tsx, and some other pages
- also fixes placeholder bg images in climate/index.tsx i think
